refactor: extract DialogOverlayBaseMixin for reusing overflow logic
Split the overflow detection (ResizeObserver, scroll and slotchange
listeners, the overflow attribute) out of DialogOverlayMixin into a
smaller DialogOverlayBaseMixin. DialogOverlayMixin now composes it and
keeps the header/footer/title renderer logic and viewport positioning.

A protected _onOverlayResize() hook lets the single resize observer stay
in the base mixin while DialogOverlayMixin layers __adjustPosition() on
top. No behavior change for vaadin-dialog.
